fail message delay & pause if dragging window
Added a delay before exiting on failure for user to view error.

An error with SDL is that a window being dragged, results in the program pausing. As the time function is independent to the program, the music rapidly speeds up to catch up with the correct timestamp.
The simplest fix was to merely check if the program was paused for longer than 0.1 seconds and if so, it should add the time it was paused onto the start time, as to continue from where it was initially paused.